test: move the AF_UNIX CONNECT hangup fixture to its own file and run it with bunRun
No-Verification-Needed: test-only change
