+++ test/js/node/http/node-http-connect.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import { bunEnv, bunExe, isLinux, nodeExe, tempDir, tls as tlsCert } from "harness";
+import { bunEnv, bunExe, bunRun, isLinux, nodeExe, tempDir, tls as tlsCert } from "harness";
 import http from "http";
 
 import { once } from "node:events";
@@ -629,50 +629,15 @@ describe("HTTP server CONNECT", () => {
     async () => {
       // An AF_UNIX peer close() on a half-open (CONNECT hand-off) socket is EPOLLHUP, which is level-triggered:
       // it must end+close the socket rather than re-dispatch forever. The server never ends its own side.
-      const fixture = /* js */ `
-        const http = require("node:http");
-        const net = require("node:net");
-        const N = 8;
-        const clients = new Map();
-        let ends = 0;
-        let closes = 0;
-        const server = http.createServer();
-        server.on("connect", (req, socket) => {
-          socket.on("error", () => {});
-          socket.on("end", () => ends++);
-          socket.on("close", () => {
-            if (++closes === N) {
-              console.log(JSON.stringify({ ends, closes }));
-              server.close();
-            }
-          });
-          // Now that the server holds the handed-off socket, hang up this
-          // request's client without it ever seeing a response.
-          clients.get(req.url).destroy();
-        });
-        server.listen(process.env.SOCK, () => {
-          for (let i = 0; i < N; i++) {
-            const target = "peer-" + i + ":443";
-            const client = net.connect(process.env.SOCK, () => {
-              client.write("CONNECT " + target + " HTTP/1.1\\r\\nHost: " + target + "\\r\\n\\r\\n");
-            });
-            client.on("error", () => {});
-            clients.set(target, client);
-          }
-        });
-      `;
       using dir = tempDir("connect-unix-hangup", {});
-      await using proc = Bun.spawn({
-        cmd: [bunExe(), "-e", fixture],
-        env: { ...bunEnv, SOCK: join(String(dir), "proxy.sock") },
-        stdout: "pipe",
-        stderr: "pipe",
+      const result = await bunRun(join(import.meta.dir, "node-http-connect-unix-hangup-fixture.js"), {
+        SOCK: join(String(dir), "proxy.sock"),
       });
-      const [stdout, stderr, exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-      expect({ stdout: stdout.trim(), stderr, exitCode }).toEqual({
+      expect(result).toEqual({
         stdout: JSON.stringify({ ends: 8, closes: 8 }),
         stderr: "",
         exitCode: 0,
+        signalCode: null,
       });
     },
   );
